Move atomic_rename from zephyr.writers to rigging.filesystem
atomic_rename (and its helpers unique_temp_path, the transient-S3 retry
logic, and _mv_with_retry) are generic filesystem utilities, not zephyr
specific. Move them down to rigging.filesystem so callers depend on the
lower layer directly instead of importing through zephyr.

- Drop the zephyr/atomic_rename_retries counter: rigging sits below zephyr
  and cannot import zephyr.counters. Retry-with-backoff behavior unchanged.
- Repoint all callsites (marin, levanter, scripts) to rigging.filesystem.
- zephyr.writers/execution now import these from rigging; remove the
  zephyr package re-export of atomic_rename.
- Move the atomic_rename / unique_temp_path tests to lib/rigging.
