+++ lib/marin/src/marin/datakit/download/formal_methods_evals.py
@@ -37,9 +37,8 @@ from dataclasses import dataclass
 from typing import Any
 
 import zstandard
-from rigging.filesystem import open_url
+from rigging.filesystem import atomic_rename, open_url
 from zephyr import Dataset, ZephyrContext
-from zephyr.writers import atomic_rename
 
 from marin.datakit.download.http_session import build_retrying_session
 from marin.datakit.ingestion_manifest import (
