+++ lib/marin/src/marin/datakit/download/uncheatable_eval.py
@@ -15,9 +15,8 @@ from dataclasses import dataclass
 from typing import Any
 
 import requests
-from rigging.filesystem import open_url
+from rigging.filesystem import atomic_rename, open_url
 from zephyr import Dataset, ZephyrContext
-from zephyr.writers import atomic_rename
 
 from marin.datakit.download.http_session import build_retrying_session
 from marin.execution import THIS_OUTPUT_PATH, ExecutorStep, VersionedValue
