+++ lib/marin/src/marin/transform/structured_text/web_data_commons.py
@@ -34,8 +34,7 @@ from marin.datakit.ingestion_manifest import (
     write_ingestion_metadata_json,
 )
 from marin.utils import fsspec_mkdirs
-from rigging.filesystem import open_url, url_to_fs
-from zephyr.writers import atomic_rename
+from rigging.filesystem import atomic_rename, open_url, url_to_fs
 
 logger = logging.getLogger(__name__)
