+++ lib/rigging/src/rigging/filesystem.py
@@ -35,6 +35,7 @@ import threading
 import time
 import urllib.error
 import urllib.request
+import uuid
 from collections.abc import Callable, Generator, Sequence
 from pathlib import PurePath
 from typing import Any
@@ -43,6 +44,7 @@ import fsspec
 from google.api_core.exceptions import Forbidden as GcpForbiddenException
 
 from rigging.distributed_lock import create_lock, default_worker_id
+from rigging.timing import ExponentialBackoff, retry_with_backoff
 
 logger = logging.getLogger(__name__)
 
@@ -773,6 +775,97 @@ def filesystem(protocol: str, **kwargs: Any) -> Any:
     if _is_gcs_protocol(protocol):
         fs = CrossRegionGuardedFS(fs)
     return fs
+
+
+# ---------------------------------------------------------------------------
+# Atomic write-and-rename
+
+
+def unique_temp_path(output_path: str) -> str:
+    """Return a unique temporary path derived from ``output_path``.
+
+    Appends ``.tmp.<uuid>`` to avoid collisions when multiple writers target the
+    same output path (e.g. during network-partition induced worker races).
+    """
+    return f"{output_path}.tmp.{uuid.uuid4().hex}"
+
+
+# AWS error codes that are safe to retry on a server-side multipart copy
+# (``s3fs.S3FileSystem.mv``). ``InvalidPart`` is the R2-specific symptom:
+# every ``UploadPartCopy`` returns 200 but ``CompleteMultipartUpload`` then
+# claims one or more parts are missing.
+_TRANSIENT_S3_ERROR_CODES = frozenset(
+    {
+        "InvalidPart",
+        "InternalError",
+        "ServiceUnavailable",
+        "SlowDown",
+        "RequestTimeout",
+        "RequestTimeTooSkewed",
+    }
+)
+
+# Fragments matched against ``str(exc)`` for the case where s3fs has already
+# translated the underlying ``botocore.ClientError`` into an ``OSError`` and
+# the structured error code is no longer reachable.
+_TRANSIENT_S3_MESSAGE_FRAGMENTS = (
+    "specified parts could not be found",
+    "InternalError",
+    "ServiceUnavailable",
+    "SlowDown",
+    "RequestTimeout",
+)
+
+
+def _is_transient_s3_error(exc: BaseException) -> bool:
+    response = getattr(exc, "response", None)
+    if isinstance(response, dict):
+        code = response.get("Error", {}).get("Code")
+        if code in _TRANSIENT_S3_ERROR_CODES:
+            return True
+    msg = str(exc)
+    return any(frag in msg for frag in _TRANSIENT_S3_MESSAGE_FRAGMENTS)
+
+
+def _mv_with_retry(fs: Any, src: str, dst: str) -> None:
+    retry_with_backoff(
+        lambda: fs.mv(src, dst, recursive=True),
+        retryable=_is_transient_s3_error,
+        max_attempts=4,
+        backoff=ExponentialBackoff(initial=1.0, maximum=8.0, factor=2.0),
+        operation=f"atomic_rename fs.mv {src} -> {dst}",
+    )
+
+
+@contextlib.contextmanager
+def atomic_rename(output_path: str, fs: Any = None) -> Generator[str, None, None]:
+    """Atomic write-and-rename via a sibling temp key.
+
+    Yields ``<output_path>.tmp.<uuid>``; on clean exit, ``fs.mv`` renames the
+    temp into the final path. On exception, the temp key is best-effort
+    deleted and the original exception re-raised.
+
+    Callers may pass a pre-constructed ``fs`` to reuse a configured
+    filesystem (e.g. an ``S3FileSystem`` with ``fixed_upload_size=True``)
+    instead of letting atomic_rename build a default one from ``output_path``.
+
+    Example:
+        with atomic_rename("output.jsonl.gz") as tmp_path:
+            write_data(tmp_path)
+        # File is now at output.jsonl.gz
+    """
+    temp_path = unique_temp_path(output_path)
+    if fs is None:
+        fs = url_to_fs(output_path)[0]
+    try:
+        yield temp_path
+        _mv_with_retry(fs, temp_path, output_path)
+    except Exception:
+        # Best-effort cleanup: temp file may not exist (writer crashed before
+        # creating it) so we tolerate any rm error and re-raise the original.
+        with contextlib.suppress(Exception):
+            fs.rm(temp_path)
+        raise
 
 
 # ---------------------------------------------------------------------------
